CHANGE: divited folder controllers   into folders and add index.js
