+++ controllers/auth/index.js
@@ -1,0 +1,13 @@
+const { HttpError, ctrlWrapper } = require("../../helpers");
+
+const register = require("./register");
+const login = require("./login");
+const logout = require("./logout");
+const getCurrent = require("./getCurrent");
+
+module.exports = {
+  register: ctrlWrapper(register),
+  login: ctrlWrapper(login),
+  getCurrent: ctrlWrapper(getCurrent),
+  logout: ctrlWrapper(logout),
+};
