+++ controllers/auth/logout.js
@@ -1,0 +1,10 @@
+const { User } = require("../../models");
+
+const logout = async (req, res) => {
+  const { _id } = req.user;
+  await User.findByIdAndUpdate(_id, { token: "" });
+
+  res.status(204).json();
+};
+
+module.exports = logout;
