--- controllers/contacts.js
@@ -1,76 +1,0 @@
-const { Contact } = require("../models");
-
-const { HttpError, ctrlWrapper } = require("../helpers");
-
-const getAll = async (req, res) => {
-  const { _id: owner } = req.user;
-  // const { page = 1, limit = 2, ...query } = req.query;
-  console.log(Contact);
-  // const skip = (page - 1) * limit;
-  // const data = await Contact.find({ owner, query }, { skip, limit }).populate("owmer", "name email")
-  const data = await Contact.find({ owner });
-  console.log(data);
-  res.json(data);
-};
-
-const getById = async (req, res) => {
-  const { contactId } = req.params;
-  const contactById = await Contact.findById(contactId);
-
-  if (!contactById) {
-    throw HttpError(404, "Not found");
-  }
-  res.json(contactById);
-};
-
-const add = async (req, res) => {
-  const { _id: owner } = req.user;
-
-  const newContact = await Contact.create({ ...req.body, owner });
-  res.status(201).json(newContact);
-};
-
-const updateById = async (req, res) => {
-  const { contactId } = req.params;
-
-  const updatedContact = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!updatedContact) {
-    throw HttpError(404, "Not found");
-  }
-  res.status(201).json(updatedContact);
-};
-const updateFavorite = async (req, res) => {
-  const { contactId } = req.params;
-
-  const updatedContactByFavorite = await Contact.findByIdAndUpdate(
-    contactId,
-    req.body,
-    {
-      new: true,
-    }
-  );
-  if (!updatedContactByFavorite) {
-    throw HttpError(404, "Not found");
-  }
-  res.json(updatedContactByFavorite);
-};
-
-const deleteById = async (req, res) => {
-  const { contactId } = req.params;
-  const removedContact = await Contact.findByIdAndRemove(contactId);
-  if (!removedContact) {
-    throw HttpError(404, "Not found");
-  }
-  res.json({ message: "Contact deleted" });
-};
-
-module.exports = {
-  getAll: ctrlWrapper(getAll),
-  getById: ctrlWrapper(getById),
-  add: ctrlWrapper(add),
-  updateById: ctrlWrapper(updateById),
-  updateFavorite: ctrlWrapper(updateFavorite),
-  deleteById: ctrlWrapper(deleteById),
-};
