+++ controllers/contacts/getById.js
@@ -1,0 +1,14 @@
+const { Contact } = require("../../models");
+const { HttpError, ctrlWrapper } = require("../../helpers");
+
+const getById = async (req, res) => {
+  const { contactId } = req.params;
+  const contactById = await Contact.findById(contactId);
+
+  if (!contactById) {
+    throw HttpError(404, "Not found");
+  }
+  res.json(contactById);
+};
+
+module.exports = getById;
