+++ controllers/index.js
@@ -1,0 +1,7 @@
+const contactsCtrl = require("../controllers/contacts");
+const authCtrl = require("../controllers/auth");
+
+module.exports = {
+  contactsCtrl,
+  authCtrl,
+};
