+++ routes/api/auth.js
@@ -2,7 +2,7 @@ const express = require("express");
 
 const router = express.Router();
 
-const ctrl = require("../../controllers/auth");
+const { authCtrl } = require("../../controllers");
 
 const { validateBody, authenticate } = require("../../middlewares");
 
@@ -11,14 +11,14 @@ const { users } = require("../../schemas");
 router.post(
   "/users/register",
   validateBody(users.registerSchema),
-  ctrl.register
+  authCtrl.register
 );
 
-router.post("/users/login", validateBody(users.loginSchema), ctrl.login);
+router.post("/users/login", validateBody(users.loginSchema), authCtrl.login);
 
-router.post("/users/logout", authenticate, ctrl.logout);
+router.post("/users/logout", authenticate, authCtrl.logout);
 
-router.get("/users/current", authenticate, ctrl.getCurrent);
+router.get("/users/current", authenticate, authCtrl.getCurrent);
 
 // router.patch("/users", ctrl.subscriptionCheck);
 module.exports = router;
