+++ routes/api/contacts.js
@@ -2,24 +2,30 @@ const express = require("express");
 
 const router = express.Router();
 
-const ctrl = require("../../controllers/contacts");
+// const ctrl = require("../../controllers/contacts");
+const { contactsCtrl } = require("../../controllers");
 
 const { validateBody, isValidId, authenticate } = require("../../middlewares");
 
 const { contacts } = require("../../schemas");
 
-router.get("/", authenticate, ctrl.getAll);
+router.get("/", authenticate, contactsCtrl.getAll);
 
-router.get("/:contactId", authenticate, isValidId, ctrl.getById);
+router.get("/:contactId", authenticate, isValidId, contactsCtrl.getById);
 
-router.post("/", authenticate, validateBody(contacts.contactsSchema), ctrl.add);
+router.post(
+  "/",
+  authenticate,
+  validateBody(contacts.contactsSchema),
+  contactsCtrl.add
+);
 
 router.put(
   "/:contactId",
   authenticate,
   isValidId,
   validateBody(contacts.contactsSchema),
-  ctrl.updateById
+  contactsCtrl.updateById
 );
 
 router.patch(
@@ -27,8 +33,8 @@ router.patch(
   authenticate,
   isValidId,
   validateBody(contacts.updateFavoriteSchema),
-  ctrl.updateFavorite
+  contactsCtrl.updateFavorite
 );
-router.delete("/:contactId", authenticate, isValidId, ctrl.deleteById);
+router.delete("/:contactId", authenticate, isValidId, contactsCtrl.deleteById);
 
 module.exports = router;
